requirements.txt
matplotlib==3.10.0
numpy==1.24.3
pandas==2.2.3
Pillow==11.1.0
scikit_learn==1.6.1
skimage==0.0
streamlit==1.41.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
utils==1.0.2
